+++ api/controllers/rate_controller.js
@@ -6,22 +6,24 @@ const Post = require('../models/posts/post_model')
 
 module.exports.add_rate = async (req, res, next) => {
     try {
-        const rate = await Rate.find({ user: req.body.user, post: req.body.post })
+        //can tai khoan de danh gia
+        const accountID = req.userData.accountId
+        const account = await Account.findById(accountID)
+        if (!account) {
+            return res.status(404).json({ error: 'account does not exist' })
+        }
+        const rate = await Rate.find({ account: accountID, post: req.body.post })
         if (rate.length > 0) {
             return res.status(409).json({ error: 'you must choose again your rate!' })
         }
-        const account = await Account.findById(req.body.account)
-        if (!account) {
-            return res.status(409).json({ error: 'user does not exist' })
-        }
         const post = await Post.findById(req.body.post)
         if (!post) {
-            return res.status(409).json({ error: 'post does not exist' })
+            return res.status(404).json({ error: 'post does not exist' })
         }
         const new_rate = new Rate({
             _id: mongoose.Types.ObjectId(),
             name: req.body.name,
-            account: req.body.account,
+            account: accountID,
             post: req.body.post,
             description: req.body.description,
             start: req.body.star,
@@ -45,6 +47,7 @@ module.exports.add_rate = async (req, res, next) => {
 
 module.exports.update_rate = async (req, res, next) => {
     try {
+        const accountID = req.userData.accountId
         const id = req.params.rateId
         const rate = await Rate.findById(id)
         if (!rate) {
@@ -55,21 +58,15 @@ module.exports.update_rate = async (req, res, next) => {
             updateOps[key] = value
         }
 
-        if (updateOps.user) {
-            const account = await Account.findById(updateOps.account)
-            if (!account) {
-                return res.status(404).json({ error: 'user does not exist' })
-            }
-        }
         if (updateOps.post) {
             const post = await Post.findById(updateOps.post)
             if (!post) {
                 return res.status(404).json({ error: 'post does not exist' })
             }
         }
-
+        
         updateOps.updated_at = new Date()
-        Rate.updateMany({ _id: id }, { $set: updateOps })
+        Rate.updateMany({ _id: id, account: accountID }, { $set: updateOps })
             .exec()
             .then(() => {
                 res.status(200).json({
@@ -88,12 +85,15 @@ module.exports.update_rate = async (req, res, next) => {
 
 module.exports.delete_rate = async (req, res, next) => {
     try {
+        const accountID = req.userData.accountId
+        // const account = await Account.findById(accountID)
+        //nguoi nao` xoa rate nguoi` do'
         const id = req.params.rateId
         const rate = await Rate.findById(id)
         if (!rate) {
             return res.status(404).json({ error: 'rate not found' })
         }
-        Rate.deleteOne({ _id: id })
+        Rate.deleteOne({ _id: id, account: accountID })
             .exec()
             .then(result => {
                 // console.log(result)
@@ -109,12 +109,14 @@ module.exports.delete_rate = async (req, res, next) => {
     }
 }
 
-module.exports.get_a_rate = (req, res, next) => {
+module.exports.get_a_rate = async (req, res, next) => {
     Rate.findById(req.params.rateId)
+        .populate({ path: 'account', select: 'username' })
+        .populate({ path: 'post', select: 'title' })
         .exec()
         .then(rate => {
             if (!rate) {
-                return res.status(200).json({
+                return res.status(404).json({
                     error: 'rate not found'
                 })
             }
@@ -127,8 +129,27 @@ module.exports.get_a_rate = (req, res, next) => {
         })
 }
 
-module.exports.get_all_rate = (req, res, next) => {
+module.exports.get_all_rate = async (req, res, next) => {
     Rate.find()
+        .populate({ path: 'account', select: 'username' })
+        .populate({ path: 'post', select: 'title' })
+        .exec()
+        .then(docs => {
+            res.status(200).json({
+                rates: docs,
+            })
+        }).catch(err => {
+            res.status(500).json({
+                error: err
+            })
+        })
+}
+//1 post thi co nhieu danh gia
+module.exports.get_all_rate_of_post = async (req, res, next) => {
+    const postId = req.params.postId
+    Rate.find({ post: postId })
+        .populate({ path: 'account', select: 'username' })
+        .populate({ path: 'post', select: 'title' })
         .exec()
         .then(docs => {
             res.status(200).json({
